Adds async example file
